added edit page and actual profile page... not backend functional yet, just frontend
