+++ packages/react-frontend/src/style/profile.module.css
@@ -1,103 +1,41 @@
-@import url('https://fonts.google.com/specimen/Josefin+Slab');
-@font-face {
-    font-family: 'Josefin Slab';
-    src: url(https://fonts.google.com/specimen/Josefin+Slab);
-  }
-  
-body {
-    background-color: #FFF3FE; /* Set the background color for the entire screen */
-    margin: 0; /* Remove default margin */
-    padding: 0; /* Remove default padding */
-  }  
-
 .background {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    min-height: 100vh;
 }
-.heading{
+.heading {
     margin-top: 30px;
     margin-bottom: 20px;
     text-align: left;
-    color:#8E5281;
-    font-family: "Josefin Slab", "san-serif";
+    color: #8E5281;
+    font-family: "Josefin Slab", sans-serif;
     font-size: 100px;
-}
-.form {
-    width: 450px;
-    height: 300px;
-    margin: 0 auto;
-    background-color: #FFF3FE;
-    justify-content: center;
-    display: contents;
   }
-.input {
-    margin-bottom: 10px;
-    margin-top: 40px;
-    width: 600px;
-    height: 50px;
-    border-radius: 3px;
-    display: flex;
-    background-color: #F9DDF3;
-    border: 1px solid #050505;
+  
+  .profileInfo {
+    text-align: left;
+    background-color: #FFF3FE;
+    padding: 20px;
+    border-radius: 10px;
     box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-    font-family: 'Josefin Slab', sans-serif;
+    width: 60%;
+    max-width: 600px;
+    margin-top: 20px;
+  }
+  
+  .info p {
     font-size: 30px;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
-
-.input::placeholder{
-    color:#8E5281;
-    font-family: 'Josefin Slab', sans-serif;
-} 
-.input:focus{
-    border-color: 1px solid #8E5281;
-    outline: 2px solid #8E5281;
-}
-.buttonContainer {
-    background-color:#8E5281;
-    color: white;
-    margin: 0 auto;
-    margin-top: 60px;
-    border:#8E5281;
-    text-align: center;
-    display:block;
-    cursor:pointer;
-    font-family: 'Josefin Slab', sans-serif;
-    font-size: 30px;
-    height: 50px;
-    width: 100px;
-}
-.buttonContainer:hover{
-    background-color: #be77af;
-}
-.newUser{
-    text-align: center;
+    color: #8E5281;
     margin-top: 50px;
-    margin-bottom: auto;
-    color:#8E5281;
-    font-family: 'Josefin Slab', sans-serif;
-    font-size: 30px;
-}
-.newUser a{
-    text-align: center;
-    color:#8E5281;
-    margin: 0 auto;
-    font-family: 'Josefin Slab', sans-serif;
-    text-decoration: underline;
-    font-size: 30px;
-
-}
-.newUser a:hover{
-    color:#be77af;
-}
-.deleteUserContainer{
-    background-color:#670b33;
-    color: white;
+    margin-bottom: 100px;
+    font-family: "Josefin Slab", sans-serif;
+  }
+  
+.editContainer{
+    background-color:rgba(170, 98, 165, 0.634);
+    color: rgb(255, 255, 255);
     margin: 0 auto;
     margin-top: 40px;
     margin-bottom: 20px;
@@ -107,13 +45,17 @@ body {
     cursor:pointer;
     height: 50px;
     width: 150px;
+    text-decoration: none;
 }
-.deleteUser{
+.edit{
     text-align: center;
     font-size: 25px;
     margin: 0% auto;
     font-family: 'Josefin Slab', sans-serif;
+    font-weight: bold;
+    text-decoration: none;
+    color: white;
 }
-.deleteUserContainer:hover{
+.editContainer:hover{
     background-color: #8E5281;
 }
